Aplicando ultimos detalles

+++ app/src/main/java/com/sagrd/travelmanagement/model/Cliente.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity
 data class Cliente(
     @PrimaryKey(autoGenerate = true)
-    val clienteId :Long,
-    var nombres: String,
-    var balance :Double
+    val clienteId : Long,
+    var nombres   : String,
+    var balance   : Float
 )
